Add "external" flag to DepositFile, change deleteAllFiles to delete non-external files only by default

+++ forms/src/main/java/cdr/forms/Deposit.java
@@ -85,9 +85,15 @@ public class Deposit {
 	}
 	
 	public void deleteAllFiles() {
+		deleteAllFiles(false);
+	}
+	
+	public void deleteAllFiles(boolean deleteExternal) {
 		for (DepositFile depositFile : this.getAllFiles()) {
-			if (depositFile.getFile() != null)
-				depositFile.getFile().delete();
+			if (depositFile.getFile() != null) {
+				if (deleteExternal || !depositFile.isExternal())
+					depositFile.getFile().delete();
+			}
 		}
 	}
